cloudsentinel-prod/backend/requirements.txt
# ── Azure ──────────────────────────────────────────────────────────────────────
azure-identity==1.17.1
azure-mgmt-costmanagement==4.0.1
azure-mgmt-resource==23.1.1
azure-mgmt-compute==30.7.0
azure-mgmt-network==25.4.0
azure-mgmt-monitor==6.0.2
azure-functions==1.20.0

# ── AWS ────────────────────────────────────────────────────────────────────────
boto3==1.34.131
botocore==1.34.131

# ── GCP ────────────────────────────────────────────────────────────────────────
google-cloud-billing==1.13.3
google-cloud-compute==1.18.2
google-cloud-monitoring==2.21.0
google-cloud-bigquery==3.23.1
google-cloud-recommender==2.15.1

# ── API / Backend ──────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
supabase==2.5.3
pandas==2.2.2
httpx==0.27.0
cryptography==42.0.8
jinja2==3.1.4
pydantic==2.7.4
python-dotenv==1.0.1
